% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurize_df.R
\name{featurize_df}
\alias{featurize_df}
\title{Featurize Data Frame}
\usage{
featurize_df(data, tsforest_obj, verbose)
}
\description{
Convert a data frame into a time series forest-ready bag of mean/sd/slope random features
}
\examples{
data("FreezerRegularTrain_TRAIN")
data("FreezerRegularTrain_TEST")
ts_obj <- new_tsforest(FreezerRegularTrain_TRAIN, target = "target")
test_df <- featurize_df(FreezerRegularTrain_TEST, tsforest_obj = ts_obj, verbose = TRUE)
}
