% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FreezerRegularTrain_TEST}
\alias{FreezerRegularTrain_TEST}
\title{Freezer Test Data
The collection of this data was part of the project titled Personalised
Retrofit Decision Support Tools for UK Homes using Smart Home Technology (REFIT).
The REFIT dataset includes data from 20 households from Loughborough area over
the period 2013-2014 (see \link{1}, \link{2}). We use data of freezers in House 1 to make
two datasets. These two datasets share a same test set and only differ in the
number of training instances. There is nothing to infer from the order of examples
in the train and test set Data created by David Murray et al. (see \link{2}). Data edited
by Shaghayegh Gharghabi, Hoang Anh Dau and Eamonn Keogh. \link{1} (Link Here) \link{2} Murray,
David et al., "A data management platform for personalised real-time energy feedback",
Proceedings of the 8th International Conference on Engery Efficiency in Domestic
Appliances and Lighting, 2015. There are two classes, one representing the power
demand of the fridge freezer in the kitchen, the other representing the power demand
of the (less frequently used) freezer in the garage. They are hard to tell apart
globally but they differ locally.}
\format{
An object of class \code{data.frame} with 2850 rows and 302 columns.
}
\usage{
FreezerRegularTrain_TEST
}
\description{
Sourced from:
http://www.timeseriesclassification.com/description.php?Dataset=FreezerRegularTrain
Freezer Test Data
}
\keyword{datasets}
