% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsforest.R
\name{intervalwise_variable_importance}
\alias{intervalwise_variable_importance}
\title{Plot Interval-Wise Variable Importance}
\usage{
intervalwise_variable_importance(
  model,
  summary_function = NULL,
  optional_example_rownumber = NULL,
  ...
)
}
\arguments{
\item{model}{a model object generated by \code{tsforest::tsforest()}}

\item{summary_function}{any summary function that returns a single value}

\item{optional_example_rownumber}{an optional rownumber of a time series from the training set}

\item{...}{any arguments to be passed to the summary function}
}
\description{
Create a ggplot summarising the variable importance over intervals
}
\details{
This function creates a ggplot that summarizes the importance of various
interval features generated by the model over the interval. It can only
be used on models where \code{ranger::ranger(..., importance = '')} has been set.
A summary function must also be passed - sum and mean seem to have the best
outcomes. Optionally, you may pass a row index to visualise an individual time
series over the top of the variable importance.
}
\examples{
data("LargeKitchenAppliances_TRAIN")
model <- tsforest::tsforest(LargeKitchenAppliances_TRAIN, verbose = FALSE, importance = 'impurity')
intervalwise_variable_importance(model, summary_function = sum)
}
