% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LargeKitchenAppliances_TEST}
\alias{LargeKitchenAppliances_TEST}
\title{Large Kitchen Appliances Test Data}
\format{
An object of class \code{data.frame} with 375 rows and 721 columns.
}
\usage{
LargeKitchenAppliances_TEST
}
\description{
These problems were taken from data recorded as part of government
sponsored study called Powering the Nation. The intention was to
collect behavioural data about how consumers use electricity
within the home to help reduce the UK's carbon footprint. The
data contains readings from 251 households, sampled in two-minute
intervals over a month. Each series is length 720 (24 hours of
readings taken every 2 minutes). Classes are Washing Machine,
Tumble Dryer and Dishwasher.
}
\details{
Sourced from:
http://www.timeseriesclassification.com/description.php?Dataset=LargeKitchenAppliances
}
\keyword{datasets}
