% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_tsforest.R
\name{new_tsforest}
\alias{new_tsforest}
\title{Create new TS Forest Object}
\usage{
new_tsforest(data, target, min_length = 2, n_trees)
}
\arguments{
\item{data}{a training data set where each row is a time series, plus a target class}

\item{target}{a character, indicating the name of the target column}
}
\description{
Helper function to create a new object of class 'tsforest'
}
\details{
Creates a new tsforest object with a training dataset and a target column. Used internally by
\code{tsforest::tsforest}, but also can be used to featurize a new data frame using \code{tsforest::featurize_df}.
}
\examples{
data("FreezerRegularTrain_TRAIN")
ts_obj <- new_tsforest(FreezerRegularTrain_TRAIN, target = "target")

}
