% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsforest.R
\name{predict.tsforest}
\alias{predict.tsforest}
\title{Predict Time Series Forest}
\usage{
\method{predict}{tsforest}(model, newdata = NULL, verbose = FALSE, ...)
}
\arguments{
\item{model}{A time series forest model object}

\item{newdata}{optional new data frame - if not supplied, will use training data}

\item{verbose}{Whether to print state of featurizing new data}

\item{...}{arguments to be passed to \code{ranger::predict.ranger}, like \code{type = "response"}}
}
\value{
Returns predictions in the form of a \code{ranger::predict.ranger} response.
}
\description{
Predict classes using a time series forest model
}
\details{
This function takes a time series forest model and returns predictions, and
featurizes new data using the same intervals as needed.
}
\examples{
\dontrun{
data("LargeKitchenAppliances_TRAIN")
data("LargeKitchenAppliances_TEST")
model <- tsforest(LargeKitchenAppliances_TRAIN, target = "target")
train_preds <- predict_tsforest(model)
test_preds <- predict_tsforest(model, newdata = LargeKitchenAppliances_TEST)
}

}
