% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsforest-package.R
\docType{package}
\name{tsforest-package}
\alias{tsforest}
\alias{tsforest-package}
\title{tsforest: Time Series Forest Classification Modelling}
\description{
This package implements Deng et al (2013)'s Time Series Classification 
    algorithm as described by timeseriesclassification.com using the ranger package
    as the random forest backend.
}
\details{
An implementation of time series forest classification models, per Deng et al (2013)
http://www.timeseriesclassification.com/algorithmdescription.php?algorithm_id=5
}
\author{
\strong{Maintainer}: Matt Simmons \email{not.given@example.com}

}
\keyword{internal}
