% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsforest.R
\name{tsforest}
\alias{tsforest}
\title{Time Series Forest}
\usage{
tsforest(df, target = "target", min_length = 2, verbose = TRUE, ...)
}
\arguments{
\item{df}{A data.frame object where each row is a time series.}

\item{target}{The name of the predicted class column, as a string}

\item{min_length}{The minimum length of the sample intervals created. Must be >= 2}

\item{verbose}{Whether or not to print the state of training as it happens.}

\item{...}{arguments to be passed to the \code{ranger::ranger()}model, like mtry}
}
\value{
Returns a time series forest model object (just a list at the moment)
}
\description{
Train a time series forest model.
}
\details{
This function trains a time series classification forest model using the methods
described by Deng et al. (2013). Specifically, the code here is based on
the description at:
http://www.timeseriesclassification.com/algorithmdescription.php?algorithm_id=5
In essence, the model splits each time series into \code{sqrt(ncol(df))} intervals,
then calcuates the mean, standard deviation and slope of those intervals, and
then uses those values for every interval as the inputs to a random forest model.
Here, we use \verb{ranger::}.
}
\examples{
\dontrun{
model <- tsforest(df)
}
}
