% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsforest-package.R
\docType{package}
\name{tsforest-package}
\alias{tsforest}
\alias{tsforest-package}
\title{tsforest: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
An implementation of time series forest classification models, per Deng et al (2013)
http://www.timeseriesclassification.com/algorithmdescription.php?algorithm_id=5
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
